Add Page Features: optional per-page storage allocations
Page Features are optional sets of storage allocated from the trailer of a disk
page.  While the current approach is handled at a cluster level, using a Page
Feature Set defined at bootstrap time, the design is granular enough that
specific features could be used on given classes of relations or even specific
relations.

The API design merges the test for a given feature with the return of the
address of the feature's storage on the page, which simplifies code that tests
and then does something with the resulting space on the page.  For instance,
with a page feature that stores 8 bytes as a page checksum, the test for such a
feature might be implemented as:

    char *extended_checksum_loc = NULL;

    /* are we using extended checksums? */
    if ((extended_checksum_loc = PageGetFeatureOffset(page, PF_EXT_CHECKSUMS)))
    {
        /* 64-bit checksum */
        page_checksum = pg_get_checksum64_page(page, (uint64*)extended_checksum_loc);
        checksum = pg_checksum64_page(page, blkno + segmentno * RELSEG_SIZE, (uint64*)extended_checksum_loc);
    }
    else
    {
        phdr = (PageHeader) page;
        page_checksum = phdr->pd_feat.checksum;
        checksum = pg_checksum_page(page, blkno + segmentno * RELSEG_SIZE);
    }

The page features are named, fixed size, and consistently offset so a cluster
that is created with the same options could use pg_upgrade for future versions,
similar to how data checksums are currently utilized.

The code overrides the interpretation of the pd_checksum field only when a
specific page flag is on the page, meaning that this code is
backwards-compatible with both checksum-enabled and -disabled historical
clusters, as well as forward-compatible with the development of additional page
features in future versions.